code/requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.0
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
deepspeed==0.14.3
distro==1.9.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.23.4
HyperPyYAML==1.2.2
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib-resources==6.1.1
importlib_metadata==7.1.0
Jinja2==3.1.3
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.1.3
lightning-utilities==0.10.1
llvmlite==0.41.1
Mako==1.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.2.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.9.0
openai-whisper==20231117
opencv-python==4.9.0.80
optuna==3.5.0
packaging==23.2
pandas==2.2.0
peft==0.3.0
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
primePy==1.3
proglog==0.1.10
protobuf==4.25.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytorch-metric-learning==2.4.1
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.4.0
scipy==1.11.4
semver==3.0.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.25
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.19.1
torch==1.13.1+cu116
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==0.13.1+cu116
torchmetrics==1.3.0.post0
torchvision==0.14.1+cu116
tqdm==4.66.1
transformers==4.41.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.3
yarl==1.9.4
youtube-dl==2021.12.17
zipp==3.17.0
